Add SLSA 0.2 converter
InTotoStatement_SlsaProvenanceZeroTwo predicate in grafeas
is wrapped as oneof predicates. That means a JSONified intoto statement
with slsa 0.2 predicate cannot be directly unmarshalled to grafeas
intoto statement.

In this commit, we add a conversion layer that converts a JSONified intoto
statement with slsa 0.2 predicate to the intoto statement compliant to
grafeas.